Capture exceptions in the C api.
Change the RtMidi*Ptr types to a simple structure.
Capture the exceptions thrown by RtMidi in the C api functions.
Use the new RtMidi*Ptr types to store the status and error message on error.

This leaves the C API mostly unchanged but allows the handling of error in
bindings to other languages (especially for dlang).