C API change in rtmidi_in_get_message
RtMidiIn::getMessage return an elapsed time in seconds. It was used to set
the size of the message array pointer.
We now use the vector size to allocate the needed memory for the message
array pointer.
We now take an extra argument: size_t * size, to give the caller the size of
the created array.