C API: allocation strategy change and documentation
In rtmidi_in_get_message, we no longer allocate the array pointed to by the
message parameter.  Instead, we assume it has already been allocated by the
caller.
This makes it easier to deal with the API from an external language (like C#)
where there is no access to the `free()` C stdlib function but still can
still pass a pointer to an array.

Added some succint documentation in rtmidi_c.h