add setClientName() functionality
The ALSA Sequencer API supports the re-setting of the client name after
the client has been created. JACK doesn't appear to support this yet
(though there way be a way using the metadata API). Other API's may
support this, but haven't been tested. This commit adds the method
setClientName(const std::string &clientName) to the RtMidi class. For
the API's for which this has not been implemented, a warning is thrown
using the RtMidiError system.