add setPortName() functionality
This commit adds a new method setPortName(const std::string &portName)
to the RtMidi class. Port renaming is implemented for the ALSA and
JACK API's. For the other API's, a warning is thrown using RtMidiError.

Note that port renaming in JACK is implemented using jack_port_rename(),
which is available on jack2 versions >= 1.9.11, and jack1 versions >=
0.125.0.