Allow WinMM users to select RT_SYSEX_BUFFER_SIZE for each port.
Quoting Christopher Arndt:
1024 is not large enough for the size of many commonly used sysex
messages e.g. for synthesizer patches.
For example, one project which uses python-rtmidi is DXConvert, which
allows to convert sysex patches for several Yamaha FM synthesizers and
receive and send them via MIDI. Yamaha sound banks for 80' FM
synthesizers usually consist of one sysex message of about 4k size. With
the default sysex buffer size, patch sending/reception would not work on
Windows with this program.
There are certainly sysex messages, that are even bigger than 8k, for
example the Korg Wavestation uses messages that can be 100k big,
IIRC. Making the sysex buffer that big for each RtMidi class instance
wouldn't make sense.

This solution removes the burden of a recompilation (in addition for
usual reasons against code duplication), and is more flexible if
different sizes are useful in the same project.

Ideally, the setting should be a parameter for openPort() but it is
specific to an API, or for the WinMMInMid constructor but it is not
public.