Fix issue with tests/sysextest.cpp not accepting key input
tests/sysextest.cpp may not accept key input
when selecting whether or not to use a virtual port,
making it impossible to make a selection.
tests/cmidiin.cpp did not have this issue.

This commit fixes tests/sysextest.cpp to prevent the issue
in the same manner as tests/cmidiin.cpp.