Add `rtmidi_set_error_callback` to C wrapper
Without setting a callback, all errors are printed to stderr, which
may not be desirable.

Rename `data` field in wrapper struct to something more descriptive.

Delete `rtmidi_error` function as the prototype mismatched the
implementation.

Templatize existing callback proxy class so we can reuse it for
error callbacks.

This is an ABI break as it adds a field to the wrapper struct.